Move `WorkflowDefinitionExporterRegressionTests` to separate test file for clarity
